+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 All Notable changes to `skyscanner-php` will be documented in this file
 
+## 2016-08-11
+- Added Browse Cache for the Flights
+
 ## 2016-07-12
 - First release with LivePricing from the Flights component of the Travel API